Changed modals delete and update to function without using portals
