+++ src/components/TabsContainter.jsx
@@ -8,26 +8,48 @@ import TabCard from './TabCard';
 import SearchBar from './SearchBar';
 import fetchSpinner from '../spinners/pacman.svg';
 
-const TabsContainer = props => {
-    return (
-        <>
-            { 
-                props.fetchLoading
-                ?   <StylesTabsContainer><img src={fetchSpinner} alt="pacman spinner" /></StylesTabsContainer>
-                :   <StylesTabsContainer>
-                        <SearchBar search={props.search} showSearchHandler={props.showSearchHandler}/>
-                        <Link to="/new" className="add-wrapper" >
-                            <span>+</span>
-                        </Link>
-                            {
-                                props.visitedTabs 
-                                ?   props.visitedTabs.map(tab => {
-                                        return( 
-                                            <Switch key={uuid()}>
-                                                <Route
-                                                    path="/home/:category"
-                                                    render={(props) => {
-                                                        if(props.match.params.category === tab.category && tab.category !== null) {
+class TabsContainer extends React.Component {
+    render() {
+        return (
+            <>
+                { 
+                    this.props.fetchLoading
+                    ?   <StylesTabsContainer><img src={fetchSpinner} alt="pacman spinner" /></StylesTabsContainer>
+                    :   <StylesTabsContainer>
+                            <SearchBar search={this.props.search} showSearchHandler={this.props.showSearchHandler}/>
+                            <Link to="/new" className="add-wrapper" >
+                                <span>+</span>
+                            </Link>
+                                {
+                                    this.props.visitedTabs 
+                                    ?   this.props.visitedTabs.map(tab => {
+                                            return( 
+                                                <Switch key={uuid()}>
+                                                    <Route
+                                                        path="/home/:category"
+                                                        render={(props) => {
+                                                            if(props.match.params.category === tab.category && tab.category !== null) {
+                                                                return <TabCard
+                                                                    {...props}
+                                                                    username={tab.username}
+                                                                    title={tab.title}
+                                                                    description={tab.description}
+                                                                    website={tab.website}
+                                                                    favicon={tab.favicon}
+                                                                    category={tab.category}
+                                                                    visited={tab.visited}
+                                                                    tabId={tab.tab_id}
+                                                                    tabs={this.props.tabs}
+                                                                    toggleModalDelete={this.props.toggleModalDelete}
+                                                                    toggleModalUpdate={this.props.toggleModalUpdate}
+                                                                    updateInfoHandler={this.props.updateInfoHandler}
+                                                                /> 
+                                                            }
+                                                        }}
+                                                    />
+                                                    <Route
+                                                        path="/home"
+                                                        render={(props) => {
                                                             return <TabCard
                                                                 {...props}
                                                                 username={tab.username}
@@ -38,38 +60,44 @@ const TabsContainer = props => {
                                                                 category={tab.category}
                                                                 visited={tab.visited}
                                                                 tabId={tab.tab_id}
-                                                                tabs={props.tabs}
+                                                                tabs={this.props.tabs}
+                                                                toggleModalDelete={this.props.toggleModalDelete}
+                                                                toggleModalUpdate={this.props.toggleModalUpdate}
+                                                                updateInfoHandler={this.props.updateInfoHandler}
                                                             /> 
-                                                        }
-                                                    }}
-                                                />
-                                                <Route
-                                                    path="/home"
-                                                    render={(props) => {
-                                                        return <TabCard
-                                                            {...props}
-                                                            username={tab.username}
-                                                            title={tab.title}
-                                                            description={tab.description}
-                                                            website={tab.website}
-                                                            favicon={tab.favicon}
-                                                            category={tab.category}
-                                                            visited={tab.visited}
-                                                            tabId={tab.tab_id}
-                                                            tabs={props.tabs}
-                                                        /> 
-                                                    }}
-                                                />
-                                            </Switch>
-                                        );            
-                                    })
-                                :   props.tabs.map(tab => {
-                                        return( 
-                                            <Switch key={uuid()}>
-                                                <Route
-                                                    path="/home/:category"
-                                                    render={(props) => {
-                                                        if(props.match.params.category === tab.category && tab.category !== null) {
+                                                        }}
+                                                    />
+                                                </Switch>
+                                            );            
+                                        })
+                                    :   this.props.tabs.map(tab => {
+                                            return( 
+                                                <Switch key={uuid()}>
+                                                    <Route
+                                                        path="/home/:category"
+                                                        render={(props) => {
+                                                            if(props.match.params.category === tab.category && tab.category !== null) {
+                                                                return <TabCard
+                                                                    {...props}
+                                                                    username={tab.username}
+                                                                    title={tab.title}
+                                                                    description={tab.description}
+                                                                    website={tab.website}
+                                                                    favicon={tab.favicon}
+                                                                    category={tab.category}
+                                                                    visited={tab.visited}
+                                                                    tabId={tab.tab_id}
+                                                                    tabs={this.props.tabs}
+                                                                    toggleModalDelete={this.props.toggleModalDelete}
+                                                                    toggleModalUpdate={this.props.toggleModalUpdate}
+                                                                    updateInfoHandler={this.props.updateInfoHandler}
+                                                                /> 
+                                                            }
+                                                        }}
+                                                    />
+                                                    <Route
+                                                        path="/home"
+                                                        render={(props) => {
                                                             return <TabCard
                                                                 {...props}
                                                                 username={tab.username}
@@ -80,36 +108,22 @@ const TabsContainer = props => {
                                                                 category={tab.category}
                                                                 visited={tab.visited}
                                                                 tabId={tab.tab_id}
-                                                                tabs={props.tabs}
+                                                                tabs={this.props.tabs}
+                                                                toggleModalDelete={this.props.toggleModalDelete}
+                                                                toggleModalUpdate={this.props.toggleModalUpdate}
+                                                                updateInfoHandler={this.props.updateInfoHandler}
                                                             /> 
-                                                        }
-                                                    }}
-                                                />
-                                                <Route
-                                                    path="/home"
-                                                    render={(props) => {
-                                                        return <TabCard
-                                                            {...props}
-                                                            username={tab.username}
-                                                            title={tab.title}
-                                                            description={tab.description}
-                                                            website={tab.website}
-                                                            favicon={tab.favicon}
-                                                            category={tab.category}
-                                                            visited={tab.visited}
-                                                            tabId={tab.tab_id}
-                                                            tabs={props.tabs}
-                                                        /> 
-                                                    }}
-                                                />
-                                            </Switch>
-                                        );            
-                                    })
-                            }
-                    </StylesTabsContainer>
-            }
-        </>
-    );
+                                                        }}
+                                                    />
+                                                </Switch>
+                                            );            
+                                        })
+                                }
+                        </StylesTabsContainer>
+                }
+            </>
+        );
+    }
 }
 
 const mapStateToProps = state => {
